[##] Refacto : Moved Siren, Siret and LuhnValidator into their own package
